Revert "Move getColorDisplayName to common"
This reverts commit 6438ac44138de275f68e7128deea0a01944145a4.
