+++ content/webapp/components/PaletteColorPicker/index.tsx
@@ -1,7 +1,10 @@
 import PaletteColorPicker, {
+  getColorDisplayName,
+  palette,
   PaletteColorPickerProps,
 } from './PaletteColorPicker';
 
+export { getColorDisplayName, palette };
 export type { PaletteColorPickerProps };
 
 export default PaletteColorPicker;
